Refact to use model/schema as in SQL alchemy instead of Pydantic model: SQLAlchemy -> model / Pydantic -> schema.
